Don't use a target feature as a key in the tree.
If a predicting feature used as a key in the tree predicting will be
broken. A data item passed into the tree for predicting has no
predicting feature. To predict it is a goal. So if the tree asking for a
predicting feature to predict it we faced a nonsense.